Refactor ProtoSerializer
